feat: refactor hud and use other fonts
